requirements.txt
python-pptx==1.0.0
ollama
